Move exec3 from cmd/state to execution
